Illo: Fix size rendering for Firefox
This update fixes the size (width) rendering of the `Illo` component for
Firefox. The solution was to add `width: 100%` for the wrapping node around
the `svg`.

The `Illo` component has also been refactored with the latest Blue conventions,
which includes namespacing and `propConnect`.
